requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PCA9685==1.0.1
Adafruit-PureIO==0.2.3
Adafruit-SSD1306==1.6.2
APScheduler==3.5.3
argh==0.26.2
docopt==0.6.2
pathtools==0.1.2
Pillow==5.3.0
pi-ina219==1.3.0
pyserial==3.4
pytz==2018.7
PyYAML==3.13
RPi.GPIO==0.6.4
six==1.11.0
spidev==3.2
stitchclient==0.6.0
tzlocal==1.5.1
watchdog==0.9.0
websockets==8.1.0
